+++ fileformats/GOB_Reader.h
@@ -4,6 +4,7 @@
 #include "../CommonTypes.h"
 #include "Archive.h"
 #include <stdio.h>
+#include <cstdint>
 
 class GOB_Reader : public Archive
 {
@@ -28,20 +29,20 @@ private:
 
 	typedef struct
 	{
-		char GOB_MAGIC[4];
-		long MASTERX;	//offset to GOX_Index_t
+        char GOB_MAGIC[4];
+		int32_t MASTERX;	//offset to GOX_Index_t
 	} GOB_Header_t;
 
 	typedef struct
 	{
-		long IX;		//offset to the start of the file.
-		long LEN;		//length of the file.
-		char NAME[13];	//file name.
+		int32_t IX;		//offset to the start of the file.
+		int32_t LEN;		//length of the file.
+        char NAME[13];	//file name.
 	} GOB_Entry_t;
 
 	typedef struct
 	{
-		long MASTERN;	//num files
+		int32_t MASTERN;	//num files
 		GOB_Entry_t *pEntries;
 	} GOB_Index_t;
